New Targeting card. Make the circle grid more generic.
