+++ client/app/cards/Pilot/PilotZoomSlider.tsx
@@ -2,23 +2,23 @@ import useAnimationFrame from "@client/hooks/useAnimationFrame";
 import { useGamepadValue } from "@client/hooks/useGamepadStore";
 import { ZoomSlider } from "@thorium/ui/Slider";
 import { useState } from "react";
-import { zoomMax, zoomMin } from "./constants";
-import { usePilotStore } from "./usePilotStore";
+import { useCircleGridStore } from "./useCircleGridStore";
 import { logslider } from "@client/utils/logSlider";
 
 export function PilotZoomSlider() {
-	const zoom = usePilotStore((store) => store.zoom);
-	const width = usePilotStore((store) => store.width);
-
+	const store = useCircleGridStore();
+	const zoom = store((store) => store.zoom);
+	const [zoomMin, zoomMax] = store((store) => [store.zoomMin, store.zoomMax]);
+	const width = store((store) => store.width);
 	// Joystick control
 	const [zoomAdjust, setZoomAdjust] = useState(0);
 	useGamepadValue("zoom-adjust", (value) => {
 		setZoomAdjust(value);
 	});
 	useGamepadValue("zoom-set", (value) => {
-		const max = width / (zoomMax * 2);
-		const min = width / (zoomMin * 2);
-		usePilotStore.setState({
+		const max = width / (zoomMax * 1.1 * 2);
+		const min = width / (zoomMin * 1.1 * 2);
+		store.setState({
 			zoom: logslider(
 				max,
 				min,
@@ -27,9 +27,9 @@ export function PilotZoomSlider() {
 		});
 	});
 	useAnimationFrame(() => {
-		const max = width / (zoomMax * 2);
-		const min = width / (zoomMin * 2);
-		usePilotStore.setState({
+		const max = width / (zoomMax * 1.1 * 2);
+		const min = width / (zoomMin * 1.1 * 2);
+		store.setState({
 			zoom: logslider(
 				max,
 				min,
@@ -45,10 +45,10 @@ export function PilotZoomSlider() {
 			<p className="text-xl">Zoom:</p>
 			<ZoomSlider
 				step={0.1}
-				zoomMin={width / (zoomMax * 2)}
-				zoomMax={width / (zoomMin * 2)}
+				zoomMin={width / (zoomMax * 1.1 * 2)}
+				zoomMax={width / (zoomMin * 1.1 * 2)}
 				value={zoom}
-				setValue={(zoom) => usePilotStore.setState({ zoom })}
+				setValue={(zoom) => store.setState({ zoom })}
 			/>
 		</>
 	);
