+++ client/app/cards/Pilot/constants.ts
@@ -1,8 +1,4 @@
-import type { Kilometer } from "@server/utils/unitTypes";
 import { Quaternion } from "three";
-
-export const zoomMax: Kilometer = 11000;
-export const zoomMin: Kilometer = 0.011;
 
 export const cameraQuaternionMultiplier = new Quaternion(
 	Math.SQRT1_2,
