+++ client/app/cards/Pilot/useCircleGridStore.tsx
@@ -4,20 +4,59 @@ import { useLiveQuery } from "@thorium/live-query/client";
 import { Matrix4, Quaternion, Vector3 } from "three";
 import create from "zustand";
 import { getWaypointRelativePosition } from "./getWaypointRelativePosition";
+import { type ReactNode, createContext, useContext, useState } from "react";
 
-export const usePilotStore = create<{
-	zoom: number;
-	tilt: number;
-	width: number;
-	height: number;
-	facingWaypoints: number[];
-}>((set) => ({
-	zoom: 100,
-	tilt: 0,
-	width: 0,
-	height: 0,
-	facingWaypoints: [],
-}));
+function createCircleGridStore({
+	zoomMin = 0.01,
+	zoomMax = 10000,
+}: { zoomMin?: number; zoomMax?: number }) {
+	return create<{
+		zoom: number;
+		zoomMin: number;
+		zoomMax: number;
+		tilt: number;
+		width: number;
+		height: number;
+		facingWaypoints: number[];
+	}>((set) => ({
+		zoom: 100,
+		zoomMin,
+		zoomMax,
+		tilt: 0,
+		width: 0,
+		height: 0,
+		facingWaypoints: [],
+	}));
+}
+
+export const CircleGirdStoreContext = createContext<ReturnType<
+	typeof createCircleGridStore
+> | null>(null);
+
+export function CircleGridStoreProvider({
+	zoomMin = 0.01,
+	zoomMax = 10000,
+	children,
+}: { zoomMin?: number; zoomMax?: number; children: ReactNode }) {
+	const [useCircleGridStore] = useState(() =>
+		createCircleGridStore({ zoomMin, zoomMax }),
+	);
+	return (
+		<CircleGirdStoreContext.Provider value={useCircleGridStore}>
+			{children}
+		</CircleGirdStoreContext.Provider>
+	);
+}
+
+export function useCircleGridStore() {
+	const store = useContext(CircleGirdStoreContext);
+	if (!store) {
+		throw new Error(
+			"useCircleGridStore must be used within a CircleGridStoreProvider",
+		);
+	}
+	return store;
+}
 
 const waypointPosition = new Vector3();
 const shipPosition = new Vector3();
@@ -28,6 +67,7 @@ const rotationQuat = new Quaternion();
 const desiredRotationQuat = new Quaternion();
 
 export function useGetFacingWaypoint() {
+	const store = useCircleGridStore();
 	const { interpolate } = useLiveQuery();
 	const [{ id, currentSystem, systemPosition }] = q.ship.player.useNetRequest();
 	const [waypoints] = q.waypoints.all.useNetRequest({ systemId: "all" });
@@ -60,6 +100,6 @@ export function useGetFacingWaypoint() {
 				facingWaypoints.push(waypoint.id);
 			}
 		}
-		usePilotStore.setState({ facingWaypoints });
+		store.setState({ facingWaypoints });
 	});
 }
