+++ client/app/cards/Targeting/data.tsx
@@ -1,0 +1,5 @@
+import { t } from "@server/init/t";
+import { pubsub } from "@server/init/pubsub";
+import { z } from "zod";
+
+export const targeting = t.router({});
