+++ client/app/cards/data.ts
@@ -6,3 +6,4 @@ export { viewscreen } from "./Viewscreen/data";
 export { alertLevel } from "./AlertLevel/data";
 export { powerGrid } from "./PowerGrid/data";
 export { remoteAccess } from "./RemoteAccess/data";
+export { targeting } from "./Targeting/data";
