+++ client/app/cards/index.ts
@@ -8,3 +8,4 @@ export * from "./Viewscreen";
 export * from "./AlertLevel";
 export * from "./PowerGrid";
 export * from "./RemoteAccess";
+export * from "./Targeting";
